+++ sdk/keyvault/azure-keyvault-keys/tests/test_examples_crypto_async.py
@@ -16,17 +16,18 @@ class TestCryptoExamples(AsyncKeyVaultTestCase):
     async def test_encrypt_decrypt_async(self, vault_client, **kwargs):
         key_client = vault_client.keys
         key_name = self.get_resource_name("crypto-test-encrypt-key")
-        key = await key_client.create_rsa_key(key_name, hsm=False)
-        client = key_client.get_cryptography_client(key)
+        key = await key_client.create_rsa_key(key_name)
+        client = vault_client.get_cryptography_client(key)
 
         # [START encrypt]
 
         from azure.keyvault.keys.crypto import EncryptionAlgorithm
 
         # encrypt returns a tuple with the ciphertext and the metadata required to decrypt it
-        key_id, algorithm, ciphertext, authentication_tag = await client.encrypt(
-            EncryptionAlgorithm.rsa_oaep, b"plaintext"
-        )
+        result = await client.encrypt(EncryptionAlgorithm.rsa_oaep, b"plaintext")
+        print(result.key_id)
+        print(result.algorithm)
+        ciphertext = result.ciphertext
 
         # [END encrypt]
 
@@ -47,8 +48,8 @@ class TestCryptoExamples(AsyncKeyVaultTestCase):
     async def test_wrap_unwrap_async(self, vault_client, **kwargs):
         key_client = vault_client.keys
         key_name = self.get_resource_name("crypto-test-wrapping-key")
-        key = await key_client.create_rsa_key(key_name, hsm=False)
-        client = key_client.get_cryptography_client(key)
+        key = await key_client.create_rsa_key(key_name)
+        client = vault_client.get_cryptography_client(key)
 
         key_bytes = b"5063e6aaa845f150200547944fd199679c98ed6f99da0a0b2dafeaf1f4684496fd532c1c229968cb9dee44957fcef7ccef59ceda0b362e56bcd78fd3faee5781c623c0bb22b35beabde0664fd30e0e824aba3dd1b0afffc4a3d955ede20cf6a854d52cfd"
 
@@ -57,14 +58,17 @@ class TestCryptoExamples(AsyncKeyVaultTestCase):
         from azure.keyvault.keys.crypto import KeyWrapAlgorithm
 
         # wrap returns a tuple with the wrapped bytes and the metadata required to unwrap the key
-        key_id, wrap_algorithm, wrapped_bytes = await client.wrap_key(KeyWrapAlgorithm.rsa_oaep, key_bytes)
+        result = await client.wrap_key(KeyWrapAlgorithm.rsa_oaep, key_bytes)
+        print(result.key_id)
+        print(result.algorithm)
+        encrypted_key = result.encrypted_key
 
         # [END wrap]
 
         # [START unwrap]
         from azure.keyvault.keys.crypto import KeyWrapAlgorithm
 
-        result = await client.unwrap_key(KeyWrapAlgorithm.rsa_oaep, wrapped_bytes)
+        result = await client.unwrap_key(KeyWrapAlgorithm.rsa_oaep, encrypted_key)
 
         # [END unwrap]
 
@@ -74,8 +78,8 @@ class TestCryptoExamples(AsyncKeyVaultTestCase):
     async def test_sign_verify_async(self, vault_client, **kwargs):
         key_client = vault_client.keys
         key_name = self.get_resource_name("crypto-test-wrapping-key")
-        key = await key_client.create_rsa_key(key_name, hsm=False)
-        client = key_client.get_cryptography_client(key)
+        key = await key_client.create_rsa_key(key_name)
+        client = vault_client.get_cryptography_client(key)
 
         # [START sign]
 
@@ -85,7 +89,10 @@ class TestCryptoExamples(AsyncKeyVaultTestCase):
         digest = hashlib.sha256(b"plaintext").digest()
 
         # sign returns a tuple with the signature and the metadata required to verify it
-        key_id, algorithm, signature = await client.sign(SignatureAlgorithm.rs256, digest)
+        result = await client.sign(SignatureAlgorithm.rs256, digest)
+        print(result.key_id)
+        print(result.algorithm)
+        signature = result.signature
 
         # [END sign]