+++ sdk/keyvault/azure-keyvault-secrets/tests/test_secrets_async.py
@@ -20,9 +20,9 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
         self.assertEqual(s1.content_type, s2.content_type)
         self.assertEqual(s1.enabled, s2.enabled)
         self.assertEqual(s1.not_before, s2.not_before)
-        self.assertEqual(s1.expires, s2.expires)
-        self.assertEqual(s1.created, s2.created)
-        self.assertEqual(s1.updated, s2.updated)
+        self.assertEqual(s1.expires_on, s2.expires_on)
+        self.assertEqual(s1.created_on, s2.created_on)
+        self.assertEqual(s1.updated_on, s2.updated_on)
         self.assertEqual(s1.recovery_level, s2.recovery_level)
         self.assertEqual(s1.key_id, s2.key_id)
 
@@ -36,7 +36,7 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
             "value should be '{}', but is '{}'".format(secret_value, secret_attributes.value),
         )
         self.assertTrue(
-            secret_attributes.properties.created and secret_attributes.properties.updated,
+            secret_attributes.properties.created_on and secret_attributes.properties.updated_on,
             "Missing required date attributes.",
         )
 
@@ -89,18 +89,18 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
             enabled = not secret.properties.enabled
             updated_secret = await client.update_secret_properties(
                 secret.name,
-                secret.properties.version,
+                version=secret.properties.version,
                 content_type=content_type,
-                expires=expires,
+                expires_on=expires,
                 tags=tags,
                 enabled=enabled,
             )
             self.assertEqual(tags, updated_secret.tags)
             self.assertEqual(secret.id, updated_secret.id)
             self.assertEqual(content_type, updated_secret.content_type)
-            self.assertEqual(expires, updated_secret.expires)
+            self.assertEqual(expires, updated_secret.expires_on)
             self.assertNotEqual(secret.properties.enabled, updated_secret.enabled)
-            self.assertNotEqual(secret.properties.updated, updated_secret.updated)
+            self.assertNotEqual(secret.properties.updated_on, updated_secret.updated_on)
             return updated_secret
 
         # update secret with version
@@ -136,7 +136,7 @@ class KeyVaultSecretTest(AsyncKeyVaultTestCase):
                 expected[secret_name] = secret
 
         # list secrets
-        result = client.list_secrets(max_results=max_secrets)
+        result = client.list_properties_of_secrets(max_results=max_secrets)
         await self._validate_secret_list(result, expected)
 
     @ResourceGroupPreparer()